+++ src/components/layout/shared/UserDropdown.tsx
@@ -125,9 +125,9 @@ const UserDropdown = () => {
                     </Typography>
                   </MenuItem>
                   {isInternalUser ? (
-                    <MenuItem className='mli-2 gap-3' onClick={e => handleDropdownClose(e, '/internal/dashboard')}>
-                      <i className='tabler-layout-dashboard' />
-                      <Typography color='text.primary'>{GH_INTERNAL_NAV.internalDashboard.label}</Typography>
+                    <MenuItem className='mli-2 gap-3' onClick={e => handleDropdownClose(e, '/admin')}>
+                      <i className='tabler-shield-lock' />
+                      <Typography color='text.primary'>{GH_INTERNAL_NAV.adminCenter.label}</Typography>
                     </MenuItem>
                   ) : null}
                   {!isInternalUser ? (